Move complex query logic out of SubscriberList model
This makes the model easier to read and understand, and isolates the
responsibility of querying for tag matches into its own class. This commit also
introduces a missing 'success case' test for `.where_tags_equal`.
